user input scanner aritmethic training

+++ java-12-operator-bitwise/src/com/tutorial/Main.java
@@ -4,13 +4,14 @@ public class Main {
     public static void main(String[] args) {
 
         // Operator Bitwise --> operator  untuk melakukan operasi pada nilai bit
-        byte a = 3;
+        int a = 3;
         byte b,c;
         String a_bits, b_bits,c_bits;
         // Operator SHIFT LEFT
         System.out.println("=======SHIFT LEFT (<<) ");
         a_bits = String.format("%8s",Integer.toBinaryString(a)).replace(" ", "0");
         System.out.printf("%s = %d\n",a_bits,a);
+
         b = (byte)(a << 3);
         b_bits = String.format("%8s",Integer.toBinaryString(b)).replace(' ', '0');
         System.out.printf("%s = %d\n",b_bits,b);
